#27 move MaterialIconColor from Core
