+++ src/MaterialIconsGenerator.Core/MaterialIconsGenerator.Core.csproj
@@ -50,7 +50,6 @@
     <Compile Include="Contracts\IProjectManager.cs" />
     <Compile Include="Contracts\IIconProvider.cs" />
     <Compile Include="Contracts\ISize.cs" />
-    <Compile Include="MaterialIconColor.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils\ColorUtils.cs" />
     <Compile Include="Utils\FileUtils.cs" />
