+++ src/MaterialIconsGenerator/MaterialIconsGenerator.csproj
@@ -71,7 +71,7 @@
     <Compile Include="Design\ViewModelDesignData.cs" />
     <Compile Include="Models\Group.cs" />
     <Compile Include="Models\Icon.cs" />
-    <Compile Include="Models\Pallete.cs" />
+    <Compile Include="Models\MaterialIconColor.cs" />
     <Compile Include="Models\Size.cs" />
     <Compile Include="Models\Type.cs" />
     <Compile Include="ViewModels\IconListViewModel.cs" />
